u-boot-iot2050: Add a temporary workaround for the SD Card issue
We have not yet reproduced this issue on an SD card, but we have
observed the QSPI issue on a basic device that only has an SD card.
Therefore, it is highly likely that the issue could also occur on an SD
card. As a precaution, we have included this workaround.

After approximately 2 months of operation, the device may fail to detect
the SD Card when rebooting. This commit provides a temporary workaround to
bypass the SD Card issue.

TODO: Implement a permanent solution for SD Card detection issues.

Related to #440.

Signed-off-by: Li Hua Qian <huaqian.li@siemens.com>