fix(packages): introduce PR in the version string for mraa
The upstream mraa project does not seem to create versions/tags for the
changes they make (version 2.2.0 was released in 2020 but the project
has continued to receive updates since). We anyhow have our own set of
patches on-top that we also need to version. Use CHANGELOG_V to include
PR into the version string of our mraa package. PR should be incremented
whenever we change the upstream revision (SRCREV) and/or update our
patches. This will help with binary distributions of the package (e.g.
for users needing to pull packages on their IOT2050 using apt).

Signed-off-by: Cedric Hombourger <cedric.hombourger@siemens.com>