linux-iot2050: Fix no display output on DisplayPort (DP).
There are two kernel issues causing no DP output:

1. For the PG2 variant, an incorrect include order in the device tree
file causes the clock for the display subsystem to be set incorrectly,
which in turn leads to a malfunction of the TI DSS (Display Subsystem).

2. The TI DSS driver uses an incorrect mode configuration for the Vout
Port, which triggers a sync-lost issue. Consequently, the TI DSS still
malfunctions even when its clock is configured correctly.

3. Workaround: TI refactored the encoder bridge chain connection logic
in v6.6-rc1, which introduced a regression affecting the IOT2050's display
output.  This commit reverts it to avoid the breakage.

Signed-off-by: Chun Jiao Zhao <chunjiao.zhao@siemens.com>