iot2050-eio-manager: Fix missing broadcast for “EIO needs update”
When the EIO controller firmware differs from the system’s expected version,
a prompt such as “Suggest to update the firmware via iot2050-eio” should be
broadcast.
Currently it isn’t, because session information cannot be obtained via psutil.
This change uses the who command to retrieve active sessions instead of psutil,
ensuring the message is broadcast as intended.

Signed-off-by: Chun Jiao Zhao <chunjiao.zhao@siemens.com>